+++ data-raw/data-raw.R
@@ -28,14 +28,9 @@ ferox$Latitude <- ferox$Latitude + as.numeric(sub("^\\w\\d\\d\\s([\\S]+).*", "\\
 ferox$Longitude <- as.numeric(sub(".*\\s\\w(\\d)\\s.*", "\\1", ferox$Position, perl = TRUE))
 ferox$Longitude <- ferox$Longitude + as.numeric(sub(".*\\s\\w\\d\\s([.\\d]+)$", "\\1", ferox$Position, perl = TRUE)) / 60
 
-pos <- SpatialPoints(cbind(ferox$Longitude * -1, ferox$Latitude), proj4string = CRS("+proj=longlat +init=epsg:27700"))
-pos %<>% spTransform(CRS = CRS("+proj=longlat +ellps=WGS84 +datum=WGS84"))
+ferox <- SpatialPointsDataFrame(cbind(ferox$Longitude * -1, ferox$Latitude), data = select(ferox, -Longitude, Latitude), proj4string = CRS("+proj=longlat +init=epsg:27700"))
+ferox %<>% spTransform(CRS = CRS("+proj=longlat +ellps=WGS84 +datum=WGS84"))
 
-ferox$Latitude <- data.frame(pos)[,2]
-ferox$Longitude <- data.frame(pos)[,1]
-
-ferox %<>% select(Date, Fish, Length, Mass, Age, Latitude, Longitude)
-
-ferox %<>% as.tbl()
+ferox@data %<>% select(Date, Fish, Length, Mass, Age)
 
 use_data(ferox, overwrite = TRUE)